Add classes to introduce the Ports and Adapters-pattern to aruba
